requirements.txt
Flask-JWT-Extended
Flask-Mail
Redis
Flask-Bcrypt
Flask
python-dotenv
Flask-CORS
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
flask-validators
imagekitio
marshmallow